fix: remove listeners during pause
Even though the _run method is a noop while the indexer is paused, there are potentially hundreds of listeners attached to the `append` and `download` events: number of feeds x number of indexes. There are 6 indexes in kappa-osm, and in a standard project of about 20 devices, that's 120 calls to _run() for every document added to the db. During a sync of 1000s that results in lots of function calls, slowing things down.